move  category management features into course folder: add CRUD operations, update models, and create UI components for category handling
