+++ lib/features/courses/domain/models/category.dart
