+++ lib/features/courses/ui/widgets/add_edit_category_dialog.dart
